Fixed compiler warnings.
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@22839 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ include/codec/ratectl.h
@@ -100,7 +100,7 @@ class OpalBitRateCalculator
 
     /** Return number of frames in history
     */
-    unsigned GetHistoryCount() const
+    size_t GetHistoryCount() const
     { return m_history.size(); }
 
     /** Return number of bytes in history
